+++ src/main/java/parkhouse/calculations/StatsSum.java
@@ -7,10 +7,10 @@ public class StatsSum extends AbstractStats {
 
 
     @Override
-    double calcExtra(IParkingController controller, double sum) {
+    long calcExtra(IParkingController controller, long sum) {
         return sum;
     }
-    void optionalPrintSthOrSo(IParkingController controller, double sum){
+    void optionalPrintSthOrSo(IParkingController controller, long sum){
 //        System.out.println("In Sum, yes in avg not used--I could print sth or give to out.println-<JS");
     }
 }
